requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\Shehab Hasnat\Desktop\lged\pyrbase\db 2\pyrebase_module.py: 1
# C:\Users\Shehab Hasnat\Desktop\lged\pyrbase\db 2\ss\ss3.py: 1
pyrebase5 == 5.0.1
